discord/discordbot/requirements.txt
langchain
openai
python-dotenv
discord.py